feat: add schemas to validate request body and refactoring some code

+++ server/routes/users.js
@@ -1,6 +1,7 @@
 "use strict";
 
 const User = require("../models/user.model");
+const { createUserSchema, loginSchema } = require("../schemas/user.schemas");
 
 module.exports = async function (fastify, opts) {
   fastify.get("/auth/users", async function (request, reply) {
@@ -13,35 +14,49 @@ module.exports = async function (fastify, opts) {
     }
   });
 
-  fastify.post("/auth/register", async function (request, reply) {
+  fastify.post(
+    "/auth/register",
+    createUserSchema,
+    async function (request, reply) {
+      try {
+        const { firstName, lastName, email, password } = request.body;
+
+        const existingUser = await User.findOne({ email });
+        if (existingUser)
+          throw fastify.httpErrors.conflict("User already registered");
+
+        const hashedPassword = await fastify.bcrypt.hash(password);
+
+        const user = await User.create({
+          firstName,
+          lastName,
+          email,
+          password: hashedPassword,
+        });
+
+        reply.status(201).send(user);
+      } catch (error) {
+        console.log(error);
+        reply.send(error);
+      }
+    }
+  );
+
+  fastify.post("/auth/login", loginSchema, async function (request, reply) {
     try {
-      const { firstName, lastName, email, password } = request.body;
+      const { email, password } = request.body;
 
-      if (!firstName)
-        throw fastify.httpErrors.badRequest("First Name Required");
+      const user = await User.findOne({ email });
+      if (!user) throw fastify.httpErrors.badRequest("User Not Registered");
 
-      if (!lastName) throw fastify.httpErrors.badRequest("Last Name Required");
+      const comparedPass = fastify.bcrypt.compare(password, user.password);
+      if (!comparedPass)
+        throw fastify.httpErrors.badRequest("Invalid Password");
 
-      if (!email) throw fastify.httpErrors.badRequest("Email Required");
+      const access_token = fastify.jwt.sign(user);
 
-      if (!password) throw fastify.httpErrors.badRequest("Password Required");
-
-      const existingUser = await User.findOne({ email });
-      if (existingUser)
-        throw fastify.httpErrors.conflict("User already registered");
-
-      const hashedPassword = await fastify.bcrypt.hash(password);
-
-      const user = await User.create({
-        firstName,
-        lastName,
-        email,
-        password: hashedPassword,
-      });
-
-      reply.status(201).send(user);
+      reply.status(200).send({ access_token });
     } catch (error) {
-      console.log(error);
       reply.send(error);
     }
   });
@@ -60,40 +75,63 @@ module.exports = async function (fastify, opts) {
     }
   });
 
-  fastify.put("/auth/user/:id", async function (request, reply) {
-    try {
-      const id = request.params.id;
-      let updatedData = { ...request.body };
+  fastify.put(
+    "/auth/user/:id",
+    { onRequest: [fastify.authenticate] },
+    async function (request, reply) {
+      try {
+        const id = request.params.id;
 
-      if (updatedData.password) {
-        updatedData.password = await fastify.bcrypt.hash(updatedData.password);
+        const userWhoLoggedIn = request.user._doc;
+
+        if (id !== userWhoLoggedIn._id.toString()) {
+          throw fastify.httpErrors.forbidden("Unauthorized to update user");
+        }
+
+        let updatedData = { ...request.body };
+
+        if (updatedData.password) {
+          updatedData.password = await fastify.bcrypt.hash(
+            updatedData.password
+          );
+        }
+
+        const user = await User.findByIdAndUpdate(id, updatedData, {
+          new: true,
+        });
+
+        if (!user) throw fastify.httpErrors.notFound("User Not Found");
+
+        reply.status(200).send(user);
+      } catch (error) {
+        reply.send(error);
       }
-
-      const user = await User.findByIdAndUpdate(id, updatedData, {
-        new: true,
-      });
-
-      if (!user) throw fastify.httpErrors.notFound("User Not Found");
-
-      reply.status(200).send(user);
-    } catch (error) {
-      reply.send(error);
     }
-  });
+  );
 
-  fastify.delete("/auth/user/:id", async function (request, reply) {
-    try {
-      const id = request.params.id;
+  fastify.delete(
+    "/auth/user/:id",
+    { onRequest: [fastify.authenticate] },
+    async function (request, reply) {
+      try {
+        const id = request.params.id;
 
-      const user = await User.findByIdAndDelete(id);
+        const userWhoLoggedIn = request.user._doc;
 
-      if (!user) throw fastify.httpErrors.notFound("User Not Found");
+        if (id !== userWhoLoggedIn._id.toString()) {
+          throw fastify.httpErrors.forbidden("Unauthorized to delete user");
+        }
 
-      reply
-        .status(200)
-        .send({ message: "User Sucessfully Deleted", user: user });
-    } catch (error) {
-      reply.send(error);
+        const user = await User.findByIdAndDelete(id);
+
+        if (!user) throw fastify.httpErrors.notFound("User Not Found");
+
+        reply
+          .status(200)
+          .send({ message: "User Sucessfully Deleted", user: user });
+      } catch (error) {
+        reply.send(error);
+      }
     }
-  });
+  );
 };
